First step in fixing the tinygettext interface.
Replaces usage of std::string as parameters and return values by const char*.
Adds a second version of POParser::parse that takes the file contents as a const char*.
Not all methods were fixed and are still exposed.
POParser::parse needs to be checked if it can actually throw an exception.